Make issue table into individual class.
